Re-work texture handling to accomodate mipmaps.

+++ src/Veldrid/Graphics/FormatHelpers.cs
@@ -19,7 +19,7 @@ namespace Veldrid.Graphics
             }
         }
 
-        public static float GetPixelSize(PixelFormat veldridFormat)
+        public static int GetPixelSize(PixelFormat veldridFormat)
         {
             switch (veldridFormat)
             {
